requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Mail==0.10.0
APScheduler==3.10.4
python-dotenv==1.0.1
